requirements.txt
# pip dependencies
# Add your dependencies here, e.g.:
torch
transformers
peft
datasets
pillow
tqdm
accelerate
